requirements.txt
Flask==2.3.3
simple-salesforce==1.12.4
python-dotenv==1.0.0 